+++ src/Rocks/Builders/Create/MockProjectedDelegateBuilder.cs
@@ -37,7 +37,8 @@ internal static class MockProjectedDelegateBuilder
 		var returnType = method.ReturnType.BuildName(typeArgumentsNamingContext);
 		var methodParameters = string.Join(", ", method.Parameters.Select(_ =>
 		{
-			var scoped = _.IsScoped ? "scoped " : string.Empty;
+			var scoped = _.IsParams ? string.Empty :
+				_.IsScoped ? "scoped " : string.Empty;
 			var direction = _.RefKind == RefKind.Ref ? "ref " : _.RefKind == RefKind.Out ? "out " : string.Empty;
 			var parameter = $"{scoped}{direction}{(_.IsParams ? "params " : string.Empty)}{_.Type.BuildName(typeArgumentsNamingContext)} @{_.Name}";
 			return $"{_.AttributesDescription}{parameter}";