+++ src/Rocks/Builders/Make/MockMethodVoidBuilder.cs
@@ -22,7 +22,8 @@ internal static class MockMethodVoidBuilder
 			var requiresNullable = _.RequiresNullableAnnotation ? "?" : string.Empty;
 			var defaultValue = _.HasExplicitDefaultValue && method.RequiresExplicitInterfaceImplementation == RequiresExplicitInterfaceImplementation.No ? 
 				$" = {_.ExplicitDefaultValue}" : string.Empty;
-			var scoped = _.IsScoped ? "scoped " : string.Empty;
+			var scoped = _.IsParams ? string.Empty :
+				_.IsScoped ? "scoped " : string.Empty;
 			var direction = _.RefKind switch
 			{
 				RefKind.Ref => "ref ",